# specs/contracts/api.yaml
openapi: 3.0.3
info:
  title: Aftertalk Core API
  description: REST API for Aftertalk Core Engine - Session management, transcription retrieval, and minutes operations
  version: 1.0.0
  contact:
    name: Aftertalk Team
servers:
  - url: https://api.aftertalk.io/v1
    description: Production server
  - url: https://api-staging.aftertalk.io/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
  /sessions:
    get:
      summary: List sessions
      description: Retrieve sessions with optional status filter and pagination
      operationId: listSessions
      tags:
        - Sessions
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SessionStatus'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionList'

    post:
      summary: Create a new session
      description: Initialize a new conversational session with participants
      operationId: createSession
      tags:
        - Sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sessions/{sessionId}:
    get:
      summary: Get session details
      description: Retrieve session information and status
      operationId: getSession
      tags:
        - Sessions
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a session
      description: Delete a session after it has ended or entered an error/completed state
      operationId: deleteSession
      tags:
        - Sessions
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Session deleted
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sessions/{sessionId}/status:
    get:
      summary: Get session status
      description: Retrieve only the current session status
      operationId: getSessionStatus
      tags:
        - Sessions
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatusResponse'
  /sessions/{sessionId}/end:
    post:
      summary: End a session
      description: Mark a session as ended and start downstream processing
      operationId: endSession
      tags:
        - Sessions
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Session ended
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /minutes:
    get:
      summary: Get minutes by session
      description: Retrieve the generated minutes document for a session
      operationId: getMinutesBySession
      tags:
        - Minutes
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session minutes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Minutes'
        '404':
          description: Minutes not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /minutes/{minutesId}:
    get:
      summary: Get minutes by ID
      description: Retrieve one minutes document
      operationId: getMinutesById
      tags:
        - Minutes
      parameters:
        - name: minutesId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Minutes document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Minutes'
        '404':
          description: Minutes not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update minutes
      description: Edit minutes (creates new version)
      operationId: updateMinutes
      tags:
        - Minutes
      parameters:
        - name: minutesId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMinutesRequest'
      responses:
        '200':
          description: Minutes updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Minutes'
        '400':
          description: Invalid minutes content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete minutes
      description: Delete a minutes document and its history
      operationId: deleteMinutes
      tags:
        - Minutes
      parameters:
        - name: minutesId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Minutes deleted
  /minutes/{minutesId}/versions:
    get:
      summary: Get minutes version history
      description: Retrieve all versions of minutes for audit trail
      operationId: getMinutesVersions
      tags:
        - Minutes
      parameters:
        - name: minutesId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Minutes version history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MinutesVersion'
  /minutes/pull/{token}:
    get:
      summary: Pull minutes with retrieval token
      description: One-shot notify_pull retrieval URL. The token in the path is the credential.
      operationId: pullMinutes
      security: []
      tags:
        - Minutes
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Minutes document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Minutes'
        '404':
          description: Invalid, expired, consumed, or unknown token
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
  schemas:
    CreateSessionRequest:
      type: object
      required:
        - participant_count
        - participants
      properties:
        participant_count:
          type: integer
          minimum: 2
        template_id:
          type: string
        metadata:
          type: string
          description: Opaque metadata string propagated to webhook payloads
        stt_profile:
          type: string
        llm_profile:
          type: string
        participants:
          type: array
          minItems: 2
          items:
            type: object
            required:
              - user_id
              - role
            properties:
              user_id:
                type: string
                maxLength: 128
              role:
                type: string
                maxLength: 64
    CreateSessionResponse:
      type: object
      required:
        - session_id
        - participants
      properties:
        session_id:
          type: string
          format: uuid
        participants:
          type: array
          items:
            type: object
            required:
              - id
              - user_id
              - role
              - token
              - expires_at
            properties:
              id:
                type: string
                format: uuid
              user_id:
                type: string
              role:
                type: string
              token:
                type: string
              expires_at:
                type: string
                format: date-time
    SessionStatus:
      type: string
      enum: [active, ended, processing, completed, error]
    Session:
      type: object
      required:
        - id
        - status
        - created_at
        - participant_count
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SessionStatus'
        created_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        participant_count:
          type: integer
          minimum: 2
        template_id:
          type: string
        metadata:
          type: string
        stt_profile:
          type: string
        llm_profile:
          type: string
    SessionList:
      type: object
      required:
        - sessions
        - total
        - limit
        - offset
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    SessionStatusResponse:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SessionStatus'
...
    Minutes:
      type: object
      required:
        - id
        - session_id
        - template_id
        - version
        - summary
        - sections
        - citations
        - status
        - provider
        - generated_at
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        template_id:
          type: string
        version:
          type: integer
        summary:
          $ref: '#/components/schemas/MinutesSummary'
        sections:
          type: object
          additionalProperties: true
          description: Template-defined section map. Values depend on each section type.
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        generated_at:
          type: string
          format: date-time
        delivered_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, ready, delivered, error]
        provider:
          type: string
    MinutesSummary:
      type: object
      required:
        - overview
        - phases
      properties:
        overview:
          type: string
        phases:
          type: array
          items:
            $ref: '#/components/schemas/MinutesPhase'
    MinutesPhase:
      type: object
      required:
        - title
        - summary
        - start_ms
        - end_ms
      properties:
        title:
          type: string
        summary:
          type: string
        start_ms:
          type: integer
        end_ms:
          type: integer
    Citation:
      type: object
      required:
        - timestamp_ms
        - text
        - role
      properties:
        timestamp_ms:
          type: integer
        text:
          type: string
        role:
          type: string
    UpdateMinutesRequest:
      type: object
      properties:
        summary:
          $ref: '#/components/schemas/MinutesSummary'
        sections:
          type: object
          additionalProperties: true
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
    MinutesVersion:
      type: object
      required:
        - id
        - minutes_id
        - version
        - content
        - edited_at
      properties:
        id:
          type: string
          format: uuid
        minutes_id:
          type: string
          format: uuid
        version:
          type: integer
        content:
          type: string
          description: JSON snapshot of previous minutes content
        edited_at:
          type: string
          format: date-time
        edited_by:
          type: string
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true